split web/api/autograd/ into more modules
